test(channels): move the UCS7604 cases beside the source they exercise
The linter is right that tests/fl/channels/ucs7604_managed.cpp mirrors no
source file. These exercise writeUCS7604, which lives in channel.cpp.hpp, so
they belong in tests/fl/channels/channel.cpp.

Moving them found a second thing: the frame case asserted gamma_2_8(127) and
got gamma_2_8(128). Neighbours in this file leave dithering enabled, and a
temporal step carries 127 to 128, so the assertion moved with the frame. It
passed in a file of its own only because nothing else ran first. encodeOnce()
now pins DISABLE_DITHER rather than inheriting whatever ran before it.
